Backend/requirements.txt
pandas
# asteroid-filterbanks==0.4.0
 ci-sdr
# configargparse==1.7
# ctc-segmentation==1.7.4
# editdistance==0.8.1
espnet-tts-frontend==0.0.3
# fast-bss-eval==0.1.3
# hydra-core==1.3.2
# jamo==0.4.1
# kaldiio==2.18.0
# librosa==0.9.2
# opt-einsum==3.3.0
# pypinyin==0.44.0
# pyworld==0.3.4
torch-complex==0.4.4
 typeguard==2.13.3
importlib-metadata==4.13.0
# numpy==1.23.5
sentencepiece==0.1.97
# humanfriendly==10.0
# pyreadline3==3.4.1
espnet_model_zoo.downloader
torch
transformers
datasets
speechbrain